changed the collision list in the collision component to a set of collision data, holding the target entity and a displacement vector
