+++ src/group4/ECS/systems/collision/CollisionData.java
@@ -1,0 +1,38 @@
+package group4.ECS.systems.collision;
+
+import com.badlogic.ashley.core.Entity;
+import group4.maths.Vector3f;
+
+public class CollisionData {
+
+    // the entity that is being collided with
+    public Entity entity;
+
+    /**
+     * the vector that needs to be added to the position entity that owns this collision data to make sure it
+     * does not intersect with entity
+     */
+    public Vector3f displacement;
+
+    /**
+     * Creates new CollisionData.
+     *
+     * @param entity       entity which collides with the owner of this collision data.
+     * @param displacement vector that needs to be added to the owners position to prevent the collision.
+     */
+    public CollisionData(Entity entity, Vector3f displacement) {
+        this.entity = entity;
+        this.displacement = displacement;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        // only allow one collision per entity
+        if (o instanceof CollisionData) {
+            CollisionData cd = (CollisionData) o;
+            return entity.equals(cd.entity);
+        }
+
+        return super.equals(o);
+    }
+}
